WIP updating pydantic config structure
