+++ agent_app_sample_code/02_agent__RAG_only_mlflow_sdk.py
@@ -131,7 +131,7 @@ experiment_info = mlflow.set_experiment(cookbook_shared_config.mlflow_experiment
 # MAGIC > *If you prefer, you can switch to using a native Python dictionary for parameterization.  Since MLflow ModelConfig only accepts YAML files or dictionaries, we dump the Pydantic model to a YAML file before passing it to MLflow ModelConfig.*
 # MAGIC
 # MAGIC You can (and often will need to) add or adjust the parameters in our template.  To add/modify/delete a parameter, you can either:
-# MAGIC 1. Modify the Pydantic classes in `agent_code_templates`
+# MAGIC 1. Modify the Pydantic classes in `utils.agents.config`
 # MAGIC 2. Create a Python dictionary in this notebook to replace the Pydantic class
 
 # COMMAND ----------
@@ -141,17 +141,17 @@ from utils.agents.config import (
     AgentConfig,
     LLMConfig,
     LLMParametersConfig,
-    RetrieverConfig,
-    RetrieverParametersConfig,
-    RetrieverSchemaConfig,
+    VectorSearchRetrieverConfig,
+    RetrieverInputSchema,
+    RetrieverOutputSchema,
 )
 import json
-from utils.agents.agent_utils import log_agent_to_mlflow
+from utils.agents import log_agent_to_mlflow
 
 # View Retriever config documentation by inspecting the docstrings
 # 
-# print(RetrieverConfig.__doc__)
-# print(RetrieverSchemaConfig.__doc__)
+# print(VectorSearchRetrieverConfig.__doc__)
+# print(RetrieverOutputSchema.__doc__)
 
 # View documentation for the parameters by inspecting the docstring
 # 
@@ -177,20 +177,20 @@ datapipeline_output_config = UnstructuredDataPipelineStorageConfig.from_yaml_fil
 #### ✅✏️ Retriever tool configuration
 ########################
 
-retriever_config = RetrieverConfig(
+retriever_config = VectorSearchRetrieverConfig(
     vector_search_index=datapipeline_output_config.vector_index,  # UC Vector Search index
     # Retriever schema, this is required by Agent Evaluation to:
     # 1. Enable the Review App to properly display retrieved chunks
     # 2. Enable metrics / LLM judges to understand which fields to use to measure the retriever
     # Each is a column name within the `vector_search_index`
-    vector_search_schema=RetrieverSchemaConfig(
+    vector_search_schema=RetrieverOutputSchema(
         primary_key="chunk_id",  # The column name in the retriever's response referred to the unique key
         chunk_text="content_chunked",  # The column name in the retriever's response that contains the returned chunk
         document_uri="doc_uri",  # The URI of the chunk - displayed as the document ID in the Review App
         additional_metadata_columns=[],  # Additional columns to return from the vector database and present to the LLM
     ),
     # Parameters defined by Vector Search docs: https://docs.databricks.com/en/generative-ai/create-query-vector-search.html#query-a-vector-search-endpoint
-    vector_search_parameters=RetrieverParametersConfig(
+    vector_search_parameters=RetrieverInputSchema(
         num_results=5,  # Number of search results that the retriever returns
         query_type="ann",  # Type of search: ann or hybrid
     ),
@@ -244,6 +244,14 @@ write_dict_to_yaml(agent_config.dict(exclude_none=True), "./configs/agent_model_
 #### Print resulting config to the console
 ########################
 print(json.dumps(agent_config.model_dump(), indent=4))
+
+# @SID what we're doing here is defining the config, 1) persisting it to a JSON model_config (why?)
+# 2) then running another notebook inline via %run (why?)
+
+# The reason we do 1) is to populate the config that's automatically loaded by step 2).
+# The reason we do 2) is to run the agent's code in the same notebook, so that we can evaluate the agent's quality, while still keeping
+# core agent logic in another notebook so that we can easily log it.
+
 
 # COMMAND ----------
