+++ agent_app_sample_code/02_agent__function_calling_mlflow_sdk.py
@@ -129,7 +129,7 @@ experiment_info = mlflow.set_experiment(cookbook_shared_config.mlflow_experiment
 # MAGIC > *If you prefer, you can switch to using a native Python dictionary for parameterization.  Since MLflow ModelConfig only accepts YAML files or dictionaries, we dump the Pydantic model to a YAML file before passing it to MLflow ModelConfig.*
 # MAGIC
 # MAGIC You can (and often will need to) add or adjust the parameters in our template.  To add/modify/delete a parameter, you can either:
-# MAGIC 1. Modify the Pydantic classes in `agent_code_templates`
+# MAGIC 1. Modify the Pydantic classes in `utils.agents.config`
 # MAGIC 2. Create a Python dictionary in this notebook to replace the Pydantic class
 
 # COMMAND ----------
@@ -139,18 +139,18 @@ from utils.agents.config import (
     AgentConfig,
     LLMConfig,
     LLMParametersConfig,
-    RetrieverConfig,
-    RetrieverParametersConfig,
-    RetrieverSchemaConfig,
+    VectorSearchRetrieverToolConfig,
+    RetrieverInputSchema,
+    RetrieverOutputSchema,
 )
-from utils.agents.agent_utils import log_agent_to_mlflow
+from utils.agents import log_agent_to_mlflow
 import json
 import yaml
 
 # # View Retriever config documentation by inspecting the docstrings
 #
-# help(RetrieverConfig)
-# help(RetrieverSchemaConfig)
+# help(VectorSearchRetrieverConfig)
+# help(RetrieverOutputSchema)
 #
 # # View documentation for the parameters by inspecting the docstring
 #
@@ -179,20 +179,20 @@ datapipeline_output_config = UnstructuredDataPipelineStorageConfig.from_yaml_fil
 # #### ✅✏️ Retriever tool that connects to the Vector Search index
 ########################
 
-retriever_config = RetrieverConfig(
+retriever_config = VectorSearchRetrieverToolConfig(
     vector_search_index=datapipeline_output_config.vector_index,  # UC Vector Search index
     # Retriever schema, this is required by Agent Evaluation to:
     # 1. Enable the Review App to properly display retrieved chunks
     # 2. Enable metrics / LLM judges to understand which fields to use to measure the retriever
     # Each is a column name within the `vector_search_index`
-    vector_search_schema=RetrieverSchemaConfig(
+    vector_search_schema=RetrieverOutputSchema(
         primary_key="chunk_id",  # The column name in the retriever's response referred to the unique key
         chunk_text="content_chunked",  # The column name in the retriever's response that contains the returned chunk
         document_uri="doc_uri",  # The URI of the chunk - displayed as the document ID in the Review App
         additional_metadata_columns=[],  # Additional columns to return from the vector database and present to the LLM
     ),
     # Parameters defined by Vector Search docs: https://docs.databricks.com/en/generative-ai/create-query-vector-search.html#query-a-vector-search-endpoint
-    vector_search_parameters=RetrieverParametersConfig(
+    vector_search_parameters=RetrieverInputSchema(
         num_results=5,  # Number of search results that the retriever returns
         query_type="ann",  # Type of search: ann or hybrid
     ),
@@ -206,6 +206,7 @@ retriever_config = RetrieverConfig(
     tool_description_prompt="Search for documents that are relevant to a user's query about the [REPLACE WITH DESCRIPTION OF YOUR DOCS].",  # the prompt used to describe when the tool so the LLM can decide when it is relevant to call.
     tool_name="retrieve_documents",  # the prompt that describes the tool's name.  Used in combination with `tool_description_prompt` to describe when the tool so the LLM can decide when it is relevant to call.
 
+    # TODO: Do we have to pass class name here?
     # Retriever internals
     tool_class_name="VectorSearchRetriever",  # Implementation detail, this is the name of the class inside the Agent's code that contains the retriever implementation.  When loading this tool, this class will be initialized.
 )
