+++ agent_app_sample_code/agents/function_calling_agent/function_calling_agent_mlflow_sdk.py
@@ -21,7 +21,7 @@ from mlflow.models import set_model, ModelConfig
 from mlflow.models.rag_signatures import StringResponse, ChatCompletionRequest, Message
 from mlflow.deployments import get_deploy_client
 import os
-from utils.agents.agent_utils import get_messages_array, extract_user_query_string, extract_chat_history
+from utils.agents.chat import get_messages_array, extract_user_query_string, extract_chat_history
 
 # COMMAND ----------
 
@@ -30,135 +30,6 @@ from utils.agents.agent_utils import get_messages_array, extract_user_query_stri
 # COMMAND ----------
 
 from databricks.vector_search.client import VectorSearchClient
-
-
-@dataclass
-class Document:
-    page_content: str
-    metadata: Dict[str, str]
-    type: str
-
-
-class VectorSearchRetriever:
-    """
-    Class using Databricks Vector Search to retrieve relevant documents.
-    """
-
-    def __init__(self, config: Dict[str, Any]):
-        self.config = config
-        self.vector_search_client = VectorSearchClient(disable_notice=True)
-        self.vector_search_index = self.vector_search_client.get_index(
-            index_name=self.config.get("vector_search_index")
-        )
-
-        vector_search_schema = self.config.get("vector_search_schema")
-        mlflow.models.set_retriever_schema(
-            primary_key=vector_search_schema.get("primary_key"),
-            text_column=vector_search_schema.get("chunk_text"),
-            doc_uri=vector_search_schema.get("document_uri"),
-        )
-
-    @mlflow.trace(span_type="TOOL", name="VectorSearchRetriever")
-    def __call__(self, query: str) -> str:
-        results = self.similarity_search(query)
-
-        context = ""
-        for result in results:
-            formatted_chunk = self.config.get("chunk_template").format(
-                chunk_text=result.get("page_content"),
-                metadata=json.dumps(result.get("metadata")),
-            )
-            context += formatted_chunk
-
-        return context.strip()
-
-    @mlflow.trace(span_type="RETRIEVER")
-    def similarity_search(
-        self, query: str, filters: Dict[Any, Any] = None
-    ) -> List[Document]:
-        """
-        Performs vector search to retrieve relevant chunks.
-
-        Args:
-            query: Search query.
-            filters: Optional filters to apply to the search, must follow the Databricks Vector Search filter spec (https://docs.databricks.com/en/generative-ai/create-query-vector-search.html#use-filters-on-queries)
-
-        Returns:
-            List of retrieved Documents.
-        """
-
-        traced_search = mlflow.trace(
-            self.vector_search_index.similarity_search,
-            name="vector_search.similarity_search",
-        )
-
-        vector_search_schema = self.config.get("vector_search_schema")
-        additional_metadata_columns = (
-            vector_search_schema.get("additional_metadata_columns") or []
-        )
-
-        columns = [
-            vector_search_schema.get("primary_key"),
-            vector_search_schema.get("chunk_text"),
-            vector_search_schema.get("document_uri"),
-        ] + additional_metadata_columns
-
-        # de-duplicate
-        columns = list(set(columns))
-
-        if filters is None:
-            results = traced_search(
-                query_text=query,
-                columns=columns,
-                **self.config.get("vector_search_parameters"),
-            )
-        else:
-            results = traced_search(
-                query_text=query,
-                filters=filters,
-                columns=columns,
-                **self.config.get("vector_search_parameters"),
-            )
-
-        # We turn the config into a dict and pass it here
-        vector_search_threshold = self.config.get("vector_search_threshold")
-        documents = self.convert_vector_search_to_documents(
-            results, vector_search_threshold
-        )
-
-        return [asdict(doc) for doc in documents]
-
-    @mlflow.trace(span_type="PARSER")
-    def convert_vector_search_to_documents(
-        self, vs_results, vector_search_threshold
-    ) -> List[Document]:
-        column_names = []
-        for column in vs_results["manifest"]["columns"]:
-            column_names.append(column)
-
-        docs = []
-        if vs_results["result"]["row_count"] > 0:
-            for item in vs_results["result"]["data_array"]:
-                metadata = {}
-                score = item[-1]
-                if score >= vector_search_threshold:
-                    metadata["similarity_score"] = score
-                    for i, field in enumerate(item[0:-1]):
-                        metadata[column_names[i]["name"]] = field
-                    # put contents of the chunk into page_content
-                    page_content = metadata[
-                        self.config.get("vector_search_schema").get("chunk_text")
-                    ]
-                    del metadata[
-                        self.config.get("vector_search_schema").get("chunk_text")
-                    ]
-
-                    doc = Document(
-                        page_content=page_content, metadata=metadata, type="Document"
-                    )
-                    docs.append(doc)
-
-        return docs
 
 # COMMAND ----------
 
@@ -193,6 +64,10 @@ class FunctionCallingAgent(mlflow.pyfunc.PythonModel):
         for tool in self.config.get("llm_config").get("tools"):
             # 1 Instantiate the tool's class w/ by passing the tool's config to it
             # 2 Store the instantiated tool to use later
+            # Should change this pattern to not use a global. But the reason we do it is because we need
+            # to JSON-serialize the configs. The goal is to have a JSON-serialized thing of tool configs
+            # and have our code be able to read it back and instantiate the tool
+            #
             self.tool_functions[tool.get("tool_name")] = globals()[tool.get("tool_class_name")](config=tool)
             self.tool_json_schemas.append(tool.get("tool_input_json_schema"))
