+++ agent_app_sample_code/utils/agents/chat.py
@@ -15,17 +15,10 @@ from utils.agents.config import (
     AgentConfig,
     LLMConfig,
     LLMParametersConfig,
-    RetrieverConfig,
-    RetrieverParametersConfig,
-    RetrieverSchemaConfig,
+    VectorSearchRetrieverConfig,
+    RetrieverInputSchema,
+    RetrieverOutputSchema,
 )
-
-@mlflow.trace(span_type="FUNCTION")
-def execute_function(tool_functions, function_name, args):
-    the_function = tool_functions[function_name]
-
-    result = the_function(**args)
-    return result
 
 def chat_completion(model_serving_client, llm_endpoint_name, llm_parameters, messages: List[Dict[str, str]]):
     traced_create = mlflow.trace(
@@ -104,50 +97,3 @@ def extract_chat_history(chat_messages_array: List[Dict[str, str]]) -> List[Dict
         raise ValueError(
             "chat_messages_array is not an Array of Dictionary, Pandas DataFrame, or array of MLflow Message."
         )
-
-def log_agent_to_mlflow(agent_config, retriever_config):
-    # Add the Databricks resources so that credentials are automatically provisioned by agents.deploy(...)
-    databricks_resources = [
-        DatabricksServingEndpoint(
-            endpoint_name=agent_config.llm_config.llm_endpoint_name
-        ),
-    ]
-
-    # Add the Databricks resources for the retriever's vector indexes
-    for tool in agent_config.llm_config.tools:
-        if type(tool) == RetrieverConfig:
-            databricks_resources.append(
-                DatabricksVectorSearchIndex(index_name=tool.vector_search_index)
-            )
-            index_embedding_model = (
-                VectorSearchClient(disable_notice=True)
-                .get_index(index_name=retriever_config.vector_search_index)
-                .describe()
-                .get("delta_sync_index_spec")
-                .get("embedding_source_columns")[0]
-                .get("embedding_model_endpoint_name")
-            )
-            if index_embedding_model is not None:
-                databricks_resources.append(
-                    DatabricksServingEndpoint(endpoint_name=index_embedding_model),
-                )
-            else:
-                raise Exception("Could not identify the embedding model endpoint resource for {tool.vector_search_index}.  Please manually add the embedding model endpoint to `databricks_resources`.")
-
-    # Specify the full path to the Agent notebook
-    model_file = "agents/function_calling_agent/function_calling_agent_mlflow_sdk"
-    model_path = os.path.join(os.getcwd(), model_file)
-
-    # Log the agent as an MLflow model
-    return mlflow.pyfunc.log_model(
-        python_model=model_path,
-        model_config=agent_config.dict(),
-        artifact_path="agent",
-        input_example=agent_config.input_example,
-        resources=databricks_resources,
-        signature=ModelSignature(
-            inputs=ChatCompletionRequest(),
-            outputs=StringResponse(), # TODO: Add in `messages` to signature
-        ),
-        code_paths=[os.path.join(os.getcwd(), "utils")],
-    )
